requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dataclasses-json==0.5.14
emoji==2.8.0
exceptiongroup==1.1.3
filetype==1.2.0
Flask==2.3.3
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==20.0.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.285
langsmith==0.0.35
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.2
pathspec==0.11.2
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
pypdf==3.16.2
pytest==7.4.2
python-dotenv==1.0.0
python-magic==0.4.27
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.20
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
unstructured==0.10.12
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
